fix-1 enhancement_338_sidebar_menu_modal
