+++ src/features/SideBarMenuModal/model/types/types.ts
@@ -1,0 +1,10 @@
+export interface IData {
+  routes?: IRoute[]
+  subtitle?: string
+  route?: string
+}
+
+export interface IRoute {
+  subtitle?: string
+  route?: string
+}
